Only delete provisions if on a build machine.
#rnx
#rb none
#jira none
#fyi jack.porter

[CL 17116882 by Ryan Durand in 4.27 branch]

+++ Engine/Source/Programs/IOS/iPhonePackager/MobileProvisionUtilities.cs
@@ -47,9 +47,12 @@ namespace iPhonePackager
             }
 
 			// Delete any provisions that exist in the folder so we don't keep old files laying around.
-			foreach (string Provision in Directory.GetFiles(LocalProvisionFolder))
+			if (Environment.GetEnvironmentVariable("IsBuildMachine") != null)
 			{
-				File.Delete(Provision);
+				foreach (string Provision in Directory.GetFiles(LocalProvisionFolder))
+				{
+					File.Delete(Provision);
+				}
 			}
 
             // copy all of the provisions from the game directory to the library
